backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pymongo==4.5.0
motor==3.3.1
python-dotenv>=1.0.0
langchain>=0.3.23
openai>=1.68.2
pydantic>=2.7.4
pydantic-settings>=2.0.3
python-multipart>=0.0.6
pypdf>=3.15.1
langsmith>=0.1.17
langchain-core>=0.3.54
langchain-text-splitters>=0.3.8
numpy>=1.26.4
thinc>=8.2.5